+++ surprise/prediction_algorithms/optimize_baselines.pyx
@@ -25,8 +25,8 @@ def baseline_als(self):
     # see also https://www.youtube.com/watch?v=gCaOa3W9kM0&t=32m55s
     # (Alex Smola on RS, ML Class 10-701)
 
-    cdef double [::1] bu = np.zeros(self.trainset.n_users)
-    cdef double [::1] bi = np.zeros(self.trainset.n_items)
+    cdef double [::1] bu = np.zeros(self.trainset.n_users, dtype=np.float64)
+    cdef double [::1] bi = np.zeros(self.trainset.n_items, dtype=np.float64)
 
     cdef int u, i
     cdef double r, err, dev_i, dev_u
@@ -63,8 +63,8 @@ def baseline_sgd(self):
         A tuple ``(bu, bi)``, which are users and items baselines.
     """
 
-    cdef double [::1] bu = np.zeros(self.trainset.n_users)
-    cdef double [::1] bi = np.zeros(self.trainset.n_items)
+    cdef double [::1] bu = np.zeros(self.trainset.n_users, dtype=np.float64)
+    cdef double [::1] bi = np.zeros(self.trainset.n_items, dtype=np.float64)
 
     cdef int u, i
     cdef double r, err
